Refactor Quote class to use String for id field
The Quote class has been refactored to use a String type for the id field instead of a Long. This change was made to align with the requirements of the MongoDB database. The ObjectIdConverter class has been added to handle the conversion between String and ObjectId types. The equals() and toString() methods have also been updated accordingly.

Fix log level for MongoDB driver in test resources

The logback.xml file in the test resources has been modified to change the log level for the org.mongodb.driver package from WARN to INFO. This change was made to provide more detailed logging information during testing.